move bn delete eddsa
