+++ js/static_dependencies/elliptic/lib/elliptic/curve/mont.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var curve = require('../curve');
-var BN = require('../BN/bn');
+var BN = require('../../BN/bn');
 var inherits = require('../../elliptic').inherits;
 var Base = curve.base;
