+++ js/static_dependencies/elliptic/lib/elliptic/ec/index.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var BN = require('../BN/bn');
+var BN = require('../../BN/bn');
 var HmacDRBG = require('../../hmac-drbg/hmac-drbg');
 var elliptic = require('../../elliptic');
 var utils = elliptic.utils;
