+++ js/static_dependencies/elliptic/lib/elliptic/ec/key.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var BN = require('../BN/bn');
+var BN = require('../../BN/bn');
 var elliptic = require('../../elliptic');
 var utils = elliptic.utils;
 var assert = utils.assert;
